Restructure http/clients a bit, adding base.py
