+++ tests/http/clients/aiohttp.py
@@ -13,7 +13,12 @@ from strawberry.types import ExecutionResult
 from tests.views.schema import Query, schema
 
 from ..context import get_context
-from . import JSON, HttpClient, Response, ResultOverrideFunction
+from .base import (
+    JSON,
+    HttpClient,
+    Response,
+    ResultOverrideFunction,
+)
 
 
 class GraphQLView(BaseGraphQLView):
