+++ tests/http/clients/async_django.py
@@ -10,7 +10,7 @@ from strawberry.types import ExecutionResult
 from tests.views.schema import Query, schema
 
 from ..context import get_context
-from . import Response, ResultOverrideFunction
+from .base import Response, ResultOverrideFunction
 from .django import DjangoHttpClient
