+++ tests/http/conftest.py
@@ -2,16 +2,18 @@ from typing import Type
 
 import pytest
 
-from .clients import HttpClient
-from .clients.aiohttp import AioHttpClient
-from .clients.asgi import AsgiHttpClient
-from .clients.async_django import AsyncDjangoHttpClient
-from .clients.async_flask import AsyncFlaskHttpClient
-from .clients.chalice import ChaliceHttpClient
-from .clients.django import DjangoHttpClient
-from .clients.fastapi import FastAPIHttpClient
-from .clients.flask import FlaskHttpClient
-from .clients.sanic import SanicHttpClient
+from .clients import (
+    AioHttpClient,
+    AsgiHttpClient,
+    AsyncDjangoHttpClient,
+    AsyncFlaskHttpClient,
+    ChaliceHttpClient,
+    DjangoHttpClient,
+    FastAPIHttpClient,
+    FlaskHttpClient,
+    HttpClient,
+    SanicHttpClient,
+)
 
 
 @pytest.fixture(
